backend/requirements.txt
flask>=3.0.0
sentence-transformers>=5.0.0
torch>=2.2.0
transformers>=4.41.0
numpy>=1.26.0
scikit-learn>=1.5.0
gunicorn>=21.2.0
uvicorn>=0.40.0
fastapi>=0.129.0
pymongo>=4.8.0
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
